# openapi.yaml
openapi: "3.1.0"
info:
  title: Nyuchi Architecture API
  version: "1.0.0"
  description: |
    Unified API for the Nyuchi Design Portal — components, brand, architecture, Ubuntu doctrine,
    3D frontend model, and design system. Canonical registry for the bundu ecosystem.

    The Nyuchi Architecture API follows resource-oriented design principles. All resources
    are served as JSON with schema.org JSON-LD metadata, CORS headers, and CDN-friendly caching.

    All data routes read from Supabase. If the database is not configured, endpoints return
    503 with setup instructions. There are no hardcoded fallbacks.

    **Base URL:** `https://design.nyuchi.com/api/v1`

    **Registry:** the stable registry across 10 architecture layers — served from the Supabase `components` table (live count via `GET /api/v1/stats` → `stable`; counts are never hardcoded in this spec so it cannot drift from the database)

    **MCP Server:** `https://design.nyuchi.com/mcp` (Streamable HTTP transport)
  contact:
    name: Nyuchi Africa
    url: https://nyuchi.com
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://design.nyuchi.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Development
...
paths:
...
  /ui/{name}:
    get:
      summary: Individual Component
      description: |
        Returns a single component with inline source code.
        Reads from Supabase. Returns 503 if database is not configured.
      operationId: getComponent
      tags: [Registry]
      parameters:
        - name: name
          in: path
          required: true
          description: Component name in kebab-case (e.g., "button", "use-toast")
          schema:
            type: string
            pattern: "^[a-z][a-z0-9-]*$"
          examples:
            ui:
              value: button
              summary: A UI component
            hook:
              value: use-toast
              summary: A hook
            lib:
              value: utils
              summary: A library utility
      responses:
        "200":
          description: Component with source code
          headers:
            Cache-Control:
              $ref: "#/components/headers/CachePublic"
            Access-Control-Allow-Origin:
              $ref: "#/components/headers/CorsWildcard"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryItem"
        "400":
          description: Invalid component name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Component not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /architecture/layers/{n}:
    get:
      summary: Per-layer detail (covenant, rules, category breakdown)
      description: |
        Single-layer detail. Wraps `get_layer_detail(p_layer_number int)`.
        Returns 400 for non-integer or out-of-range `n` (must be 1-10).
      operationId: getLayerDetail
      tags: [Architecture]
      parameters:
        - name: n
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 10
      responses:
        "200":
          description: Layer detail with covenant, rules, and category breakdown.
          headers:
            Cache-Control:
              $ref: "#/components/headers/CachePublic"
            Access-Control-Allow-Origin:
              $ref: "#/components/headers/CorsWildcard"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: object }
        "400":
          description: Invalid layer number.
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  headers:
    CachePublic:
      schema:
        type: string
        example: "public, max-age=3600, s-maxage=86400"
    CorsWildcard:
      schema:
        type: string
        example: "*"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ServiceUnavailableError:
      type: object
      required: [error, setup]
      properties:
        error:
          type: string
          example: "Database not configured"
        setup:
          type: string
          description: Instructions for configuring Supabase
          example: "Set NEXT_PUBLIC_SUPABASE_URL and NEXT_PUBLIC_SUPABASE_ANON_KEY environment variables"

    # ── Discovery ──────────────────────────────────────────────────
...
    RegistryItem:
      type: object
      required: [$schema, name, type, description, dependencies, registryDependencies, files]
      properties:
        $schema:
          type: string
          const: "https://ui.shadcn.com/schema/registry-item.json"
        name:
          type: string
        type:
          type: string
          enum: ["registry:ui", "registry:hook", "registry:lib"]
        description:
          type: string
        dependencies:
          type: array
          items:
            type: string
        registryDependencies:
          type: array
          items:
            type: string
        files:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RegistryFile"
    RegistryFile:
      type: object
      required: [path, type, content]
      properties:
        path:
          type: string
          example: "components/ui/button.tsx"
        type:
          type: string
          example: "registry:ui"
        content:
          type: string
          description: Full source code of the component file

    # ── Architecture ───────────────────────────────────────────────
...
  responses:
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: Supabase not configured — returns setup instructions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceUnavailableError"